+++ app/dependency_file_fetchers/ruby_dependency_file_fetcher.rb
@@ -1,5 +1,4 @@
-require "bumper/dependency_file"
-require "bumper/workers"
+require "./app/dependency_file"
 require "github"
 
 module DependencyFileFetchers
