+++ app/dependency_file_parsers/ruby_dependency_file_parser.rb
@@ -1,5 +1,5 @@
 require "gemnasium/parser"
-require "bumper/dependency"
+require "./app/dependency"
 
 module DependencyFileParsers
   class RubyDependencyFileParser
