+++ app/workers/dependency_file_parser.rb
@@ -1,10 +1,9 @@
 require "hutch"
-$LOAD_PATH << "lib"
+require "./app/boot"
+require "./app/dependency_file"
+require "./app/dependency_file_parsers/ruby_dependency_file_parser"
 
 $stdout.sync = true
-
-require "bumper/dependency_file"
-require "bumper/dependency_file_parsers/ruby_dependency_file_parser"
 
 module Workers
   class DependencyFileParser
