--- lib/bumper/workers.rb
@@ -1,7 +1,0 @@
-module Workers
-  require "bumper/workers/dependency_file_fetcher"
-  require "bumper/workers/dependency_file_parser"
-  require "bumper/workers/update_checker"
-  require "bumper/workers/dependency_file_updater"
-  require "bumper/workers/pull_request_creator"
-end
