+++ lib/github.rb
@@ -1,4 +1,3 @@
-require "bumper/boot"
 require "octokit"
 
 module Github
