+++ spec/bumper/dependency_file_fetchers/ruby_dependency_file_fetcher_spec.rb
@@ -1,6 +1,6 @@
 require "spec_helper"
-require "bumper/dependency_file"
-require "bumper/dependency_file_fetchers/ruby_dependency_file_fetcher"
+require "./app/dependency_file"
+require "./app/dependency_file_fetchers/ruby_dependency_file_fetcher"
 
 RSpec.describe DependencyFileFetchers::RubyDependencyFileFetcher do
   let(:file_fetcher) { described_class.new(repo) }
